Move classes to main/java
