+++ src/main/java/player_properties.txt
